requirements.txt
ipython
pillow
numpy
torch
requests
tqdm
matplotlib

